+++ controls/views.py
@@ -38,12 +38,14 @@ from urllib.parse import quote
 
 from api.siteapp.serializers.tags import SimpleTagSerializer
 from guidedmodules.models import Task, Module, AppVersion, AppSource
+from guidedmodules.app_loading import ModuleDefinitionError
 from siteapp.model_mixins.tags import TagView
 from simple_history.utils import update_change_reason
 
-from siteapp.models import Project, Organization, Tag, User
+from siteapp.models import Project, Organization, Folder, Portfolio, Tag, User, Role, Party, Appointment, Request, Proposal
 from siteapp.settings import GOVREADY_URL
-from siteapp.utils.views_helper import project_context
+from siteapp.utils.views_helper import project_context, start_app, get_compliance_apps_catalog, \
+    get_compliance_apps_catalog_for_user, get_compliance_apps_catalog_for_user
 from system_settings.models import SystemSettings
 from .forms import ElementEditForm, ElementEditAccessManagementForm
 from .forms import ImportOSCALComponentForm, SystemAssessmentResultForm
@@ -51,7 +53,6 @@ from .forms import StatementPoamForm, PoamForm, ElementForm, DeploymentForm, Sta
 from .models import *
 from .utilities import *
 from siteapp.utils.views_helper import project_context
-from siteapp.models import Role, Party, Appointment, Request, Proposal
 from integrations.models import Integration
 
 logging.basicConfig()
@@ -3795,8 +3796,37 @@ def get_integrations_system_info(request, system_id):
             status = f"{csam_data.get('operationalStatus', 'Missing')}"
             impact = f"{csam_data.get('categorization', 'Missing')}"
         else:
-            # retreive fresh data from CSAM
-            pass
+            csam_data = {'id': 111,
+             'externalId': 'string',
+             'name': 'System A',
+             'description': 'This is a simple test system',
+             'acronym': 'string',
+             'organization': 'string',
+             'subOrganization': 'string',
+             'operationalStatus': 'string',
+             'systemType': 'string',
+             'financialSystem': 'string',
+             'classification': 'string',
+             'contractorSystem': True,
+             'fismaReportable': True,
+             'criticalInfrastructure': True,
+             'missionCritical': True,
+             'purpose': 'string',
+             'ombExhibit': 'string',
+             'uiiCode': 'string',
+             'investmentName': 'string',
+             'portfolio': 'string',
+             'priorFyFunding': 0,
+             'currentFyFunding': 0,
+             'nextFyFunding': 0,
+             'categorization': 'string',
+             'fundingImportStatus': 'string'}
+            purpose = f"{csam_data.get('purpose_short', 'Missing')}"
+            organization_name = f"{csam_data.get('organization', 'Missing')} {csam_data.get('subOrganization', 'Missing')}"
+            other_id = f"{csam_data.get('id', 'Missing')}"
+            system_type = f"{csam_data.get('systemType', 'Missing')}"
+            status = f"{csam_data.get('operationalStatus', 'Missing')}"
+            impact = f"{csam_data.get('categorization', 'Missing')}"
     else:
         # Couldn't find CSAM system
         # Sample systems from DHS SORNs
@@ -3915,6 +3945,98 @@ def get_integrations_system_events(request, system_id):
         { "event_tag": "SYS", "event_summary": "ISSO appointed - Janice Avery (contracor) has been appointed as ISSO for System"}
     ]
     return system_events
+
+
+@login_required
+def create_system_from_string(request):
+    """Create a system in GovReady-Q based on info from a URL"""
+
+    # communication = set_integration()
+    print("request", request.GET)
+    new_system_str = request.GET.get("s", None)
+    new_system_name = new_system_str.replace("https://", "").replace("http://", "")
+    new_system_description = f"System created from url."
+
+    # Examine remote site for information
+    # Handle error case of no URL
+
+    # Check if system aleady exists with domain name
+    # if not System.objects.filter(Q(info__contains={"csam_system_id": csam_system_id})).exists():
+    if True:
+        # Create new system
+        # What is default template?
+        source_slug = "govready-q-files-startpack"
+        app_name = "speedyssp"
+        # can user start the app?
+        # Is this a module the user has access to? The app store
+        # does some authz based on the organization.
+        catalog = get_compliance_apps_catalog_for_user(request.user)
+        for app_catalog_info in catalog:
+            if app_catalog_info["key"] == source_slug + "/" + app_name:
+                # We found it.
+                break
+        else:
+            raise Http404()
+        # Start the most recent version of the app.
+        appver = app_catalog_info["versions"][0]
+        organization = Organization.objects.first()  # temporary
+        default_folder_name = "Started Apps"
+        folder = Folder.objects.filter(
+            organization=organization,
+            admin_users=request.user,
+            title=default_folder_name,
+        ).first()
+        if not folder:
+            folder = Folder.objects.create(organization=organization, title=default_folder_name)
+            folder.admin_users.add(request.user)
+        task = None
+        q = None
+        # Get portfolio project should be included in.
+        if request.GET.get("portfolio"):
+            portfolio = Portfolio.objects.get(id=request.GET.get("portfolio"))
+        else:
+            if not request.user.default_portfolio:
+                request.user.create_default_portfolio_if_missing()
+            portfolio = request.user.default_portfolio
+        # import ipdb; ipdb.set_trace()
+        try:
+            project = start_app(appver, organization, request.user, folder, task, q, portfolio)
+        except ModuleDefinitionError as e:
+            error = str(e)
+        # Associate System with CSAM system
+        new_system = project.system 
+        new_system.info = {
+            "created_from_input": new_system_str,
+            "system_description": new_system_description
+        }
+        new_system.save()
+        # Update System name to URL system name
+        nsre = new_system.root_element
+        # Make sure system root element name is unique
+        name_suffix = ""
+        while Element.objects.filter(name=f"{new_system_name}{name_suffix}").exists():
+            # Element exists with that name
+            if name_suffix == "":
+                name_suffix = 1
+            else:
+                name_suffix = str(int(name_suffix)+1)
+        if name_suffix == "":
+            nsre.name = new_system_name
+        else:
+            nsre.name = f"{new_system_name}{name_suffix}"
+        nsre.save()
+        # Update System Project title to CSAM system name
+        prt = project.root_task
+        prt.title_override = nsre.name
+        prt.save()
+        logger.info(event=f"create_system_from_url url {new_system_name}",
+                object={"object": "system", "id": new_system.id},
+                user={"id": request.user.id, "username": request.user.username})
+        messages.add_message(request, messages.INFO, f"Created new System in GovReady based on URL {new_system_name}.")
+
+        # Redirect to the new system/project.
+        return HttpResponseRedirect(project.get_absolute_url())   
+        # return HttpResponseRedirect(f"/system/{new_system.id}/aspen/summary")
 
 @login_required
 def system_summary_1_aspen(request, system_id):