+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Release adds support for private components and integrations with third party se
 * Support navigating to specific tab on component library component page using URL hash (#) reference.
 * Protype integrations System Summary page.
 * Refactor and OIDC authentication for proper testing of admin and not admin roles.
+* Create a new system via name given by a string in URL.
 * Add a large set of sample components (150+) generated from STIGs.
 
 **Bug fixes**